Move the telemetry constants to the util package
Sits alongside EventUtil and MetaUtil in internal.utils, named for the same
convention. Its consumers are the invoker and the rpc proxy, neither of which
owns it.
